requirements.txt
#pip install torch==1.9.1+cu111 torchvision==0.10.1+cu111 torchaudio==0.9.1 -f https://download.pytorch.org/whl/torch_stable.html
tqdm
transformers
datasets
umap-learn
umap-learn[plot]